NeighborlyAPI/requirements.txt
azure-functions==1.9.0
azure-cosmos==4.2.0
azure-cli-core==2.34.1
bson==0.5.10
pymongo==4.2.0
Flask==2.2.2
requests==2.28.1